Fix #6
A macro name was renamed, and was not reflected in
the lowfat.h file.  This lead to lowfat_base()
using the non-pow2 definition in pow2 mode, which
broke lowfat_malloc().  This only affected the
runtime, not instrumentation.

A second problem was that the Test.cpp suite was
assuming the allocators work correctly, so did not
detect this bug.  This has also been corrected.